Struct owned_fd::FdRef

pub struct FdRef {
    // some fields omitted
}

A zero-cost (well, very, very, low cost) borrow of an OwnedFd.

This cannot be constructed directly, and can only exist as &FdRef.

As a result, it might be slightly larger than a bare RawFd.

Methods

impl FdRef

unsafe fn from_unowned_raw<'a>(i: RawFd) -> &'a FdRef

Construct a FdRef reference from a RawFd. No ownership is taken.

unsafety:

@i must be a valid fd

the lifetime 'a must be appropriately bound