Struct owned_fd::OwnedFd

pub struct OwnedFd {
    // some fields omitted
}

OwnedFd is an RAII wrapper around RawFd: it automatically closes on Drop & provides Borrow::borrow() functionality to support lifetimes & borrow checking around the use of file descriptors

Compared to using File, TcpStream, etc as a wrapper to close on drop, OwnedFd:

allows any type of filedescriptor to be used safely

has no overhead greater than a RawFd (no buffer, metadata, or other allocations)

allows use of the borrow system to ensure drop (close) happens only when all users of an ownedfd have released it.

Methods

impl OwnedFd

unsafe fn from_unowned_raw(i: RawFd) -> Result<OwnedFd>

Given a raw file descriptor that may be owned by another (ie: another data structure might close it), create an owned version that we have control over (via OwnedF::dup())

For taking ownership, see FromRawFd::from_raw_fd().

Unsafety:

i must be a valid file descriptor (of any kind)

fn dup(&self) -> Result<OwnedFd>

Duplicate this OwnedFd, and allow the error to be detected.

Clone::clone() uses this, but panics on error

fn from<T: IntoRawFd>(i: T) -> Self

Given a type that impliments IntoRawFd construct an OwnedFd.

This is safe because we assume the promises of IntoRawFd are followed.

NOTE: ideally, we'd impl this as From<T>, but current rust doesn't allow that. Revisit when specialization stabilizes.